Move plex_server logic to separate module
This removes it's dependency from main module.
